SpreadsheetViewportNavigationContextTesting2 extracted SpreadsheetViewportNavigationContextTesting
